backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
python-whois
gunicorn
joblib